feat: remove unused response and refactor
