+++ src/pages/api/repair-request/[id].ts
@@ -5,7 +5,6 @@ import { HttpStatusCode } from "axios";
 import apiHandler from "@/lib/api-handler";
 import prisma from "@/lib/prisma";
 import { updateRepairRequestSchema } from "@/schema/repair-request";
-import repairRequestService from "@/services/repairRequest";
 import { RepairRequestResponse } from "@/types";
 
 export default apiHandler({
@@ -39,7 +38,7 @@ async function updateRepairRequest(
     );
   }
 
-  const repairAttempt = await prisma.repairRequest.update({
+  await prisma.repairRequest.update({
     where: { id: id as string },
     data: {
       itemMaterial: itemMaterial,
@@ -50,9 +49,5 @@ async function updateRepairRequest(
     }
   });
 
-  const repairRequestResponse = (
-    await repairRequestService.toClientResponse([repairAttempt])
-  )[0];
-
-  return res.status(200).json(repairRequestResponse);
+  return res.status(204);
 }
